refactor: move error modal to feddback page.
