+++ src/pages/main-page/MainFooter/MainFooter.tsx
@@ -1,57 +1,17 @@
 import { AndroidFilled, AppleFilled } from "@ant-design/icons"
-import { ErrorStatus500 } from "@components/ModalWindows/FeedbackModal";
 import { useAppDispatch } from "@hooks/typed-react-redux-hooks";
-import { useModalWindow } from "@hooks/use-modal-windows";
-import { IErrorResponse } from "@interfaces/response-error.interface";
-import { appActions } from "@redux/app.slice";
-import { authActions } from "@redux/auth.slice";
-import { FetchBaseQueryError } from "@reduxjs/toolkit/query";
 import { Paths } from "@routes/index";
-import { useGetReviewsQuery } from "@services/feedback-service";
 import { Button, Card, Grid } from "antd"
 import { Footer } from "antd/lib/layout/layout"
-import { useEffect, useState } from "react";
 import { push } from "redux-first-history";
 
 
-type CustomError = FetchBaseQueryError & IErrorResponse
+
 
 export const MainFooter: React.FC = () => {
   const { useBreakpoint } = Grid;
   const screens = useBreakpoint();
   const dispatch = useAppDispatch();
-  const { isModalOpen, showModal, handleCancel } = useModalWindow();
-  const [reviewButtonClicked, setReviewButtonClicked] = useState<boolean>(false);
-  const {
-    isLoading,
-    isSuccess,
-    isError,
-    error
-  } = useGetReviewsQuery(reviewButtonClicked);
-
-  const getFeedback = () => {
-    setReviewButtonClicked(true);
-  }
-  useEffect(() => {
-    dispatch(appActions.setIsLoading(isLoading));
-  }, [isLoading, dispatch]);
-  useEffect(() => {
-    if (isSuccess)
-      dispatch(push(Paths.Feedbacks));
-    if (isError && error) {
-      const customError = error as CustomError;
-
-      if (customError.data.statusCode === 403) {
-        localStorage.removeItem('access_token');
-        dispatch(authActions.setAccessToken(''))
-        dispatch(push(Paths.Auth.Login));
-        return
-      }
-      if (customError.data.statusCode !== 403)
-        showModal()
-    }
-  }, [isSuccess, isError, dispatch])
-
 
 
   return (
@@ -67,7 +27,7 @@ export const MainFooter: React.FC = () => {
 
       <Button
         data-test-id='see-reviews'
-        onClick={getFeedback}
+        onClick={() => dispatch(push(Paths.Feedbacks))}
         type='link'>
         Смотреть отзывы
       </Button>
@@ -101,9 +61,6 @@ export const MainFooter: React.FC = () => {
           IOS
         </Button>
       </Card>
-      <ErrorStatus500
-        open={isModalOpen}
-        onCancel={handleCancel} />
     </Footer>
   )
 }
